requirements.txt
streamlit>=1.27.0
pandas>=1.5.0
numpy>=1.21.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
scikit-learn>=1.1.0
umap-learn>=0.5.0
hdbscan>=0.8.0
altair>=4.0.0
openpyxl>=3.0.0